auth solution

+++ src/types/product/product.resolvers.js
@@ -11,24 +11,40 @@ const productsTypeMatcher = {
 
 /** product */
 const product = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError()
+  }
+
   return Product.findById(args.id)
     .lean()
     .exec()
 }
 
 const newProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError()
+  }
+
   // use this fake ID for createdBy for now until we talk auth
-  const createdBy = mongoose.Types.ObjectId()
+  const createdBy = ctx.user._id
   return Product.create({ ...args.input, createdBy })
 }
 
 const products = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError()
+  }
+
   return Product.find({})
     .lean()
     .exec()
 }
 
 const updateProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError()
+  }
+
   const update = args.input
   return Product.findByIdAndUpdate(args.id, update, { new: true })
     .lean()
@@ -36,6 +52,10 @@ const updateProduct = (_, args, ctx) => {
 }
 
 const removeProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError()
+  }
+
   return Product.findByIdAndRemove(args.id)
     .lean()
     .exec()
